add init, add check feasibility

+++ cvrpptpl/include/problem/customer.h
@@ -10,5 +10,5 @@ struct Customer : Node {
   bool isSelfPickup;
   bool isFlexible;
   std::vector<int> lockerPreferenceIdxs;
-  Customer(int index, float x, float y, int serviceTime, int demand, bool isSelfPickup, bool isFlexible, const std::vector<int> & lockerPreferenceIdxs);
+  Customer(int index, double x, double y, int serviceTime, int demand, bool isSelfPickup, bool isFlexible, const std::vector<int> & lockerPreferenceIdxs);
 };
