+++ cvrpptpl/include/problem/locker.h
@@ -5,5 +5,5 @@
 struct Locker : Node {
   int serviceTime;
   int capacity;
-  Locker(int index, float x, float y, int serviceTime, int capacity);
+  Locker(int index, double x, double y, int serviceTime, int capacity);
 };
