+++ cvrpptpl/include/problem/mrt_line.h
@@ -6,9 +6,9 @@ struct MrtLine {
   const Locker startStation;
   const Locker endStation;
   const int startStationServiceTime;
-  const float cost;
+  const double cost;
   const int freightCapacity;
 
   MrtLine(const Locker& startStation, const Locker& endStation,
-    int startStationServiceTime, float cost, int freightCapacity);
+    int startStationServiceTime, double cost, int freightCapacity);
 };
