+++ cvrpptpl/src/problem/customer.cpp
@@ -1,7 +1,7 @@
 #include "problem/customer.h"
 #include "problem/node.h"
 
-Customer::Customer(int index, float x, float y, int serviceTime, int demand, 
+Customer::Customer(int index, double x, double y, int serviceTime, int demand, 
                     bool isSelfPickup, bool isFlexible, const std::vector<int> & lockerPreferenceIdxs)
   : Node(index, x, y), 
   serviceTime(serviceTime),
