+++ cvrpptpl/src/problem/locker.cpp
@@ -1,7 +1,7 @@
 #include "problem/locker.h"
 #include "problem/node.h"
 
-Locker::Locker(int index, float x, float y,
+Locker::Locker(int index, double x, double y,
   int serviceTime, int capacity)
   : Node(index, x, y),
   serviceTime(serviceTime),
