+++ cvrpptpl/src/problem/node.cpp
@@ -1,3 +1,3 @@
 #include "problem/node.h"
 
-Node::Node(int index, float x, float y) : idx(index), coord({x,y}){}
+Node::Node(int index, double x, double y) : idx(index), coord({x,y}){}
